LayoutTests/fast/dom/Element/documentElement-scrollTop.html

<!DOCTYPE html>
<html>
    <head>
        <style>
            div {
                height: 9999px;
                width:9999px;
            }
        </style>
        <script src="../../../resources/js-test-pre.js"></script>
        <script>
            function runTest() {
                description('Tests that for standard mode document.documentElement.scrollTop returns the scroll top value');

                document.documentElement.scrollTop = 500;
                shouldBe("document.documentElement.scrollTop","0");
                isSuccessfullyParsed();
            }
        </script>
    </head>
    <body onload="runTest()">
        <div></div>
    </body>
</html>
